fix(tests): move the cross-boundary rubric guards out of the pack
CI's `pack-tests-stay-in-pack` check caught what local pytest and lint could
not: two guards added this session read `docs/product/briefs/` from inside
`packs/core/tests/`, reaching above the pack via `PACK_ROOT.parents[1]`. Pack
tests must anchor at their owning pack; repository-level coverage belongs in
`tests/conformance` or `tests/roster`.

Both guards relate a shipped pack file to a brief, so they are repository-level
by nature and now live in
`tests/roster/test_spec_authoring_rubric_brief_boundary.py` with their
calibration note. The pack test keeps every pack-local criterion-shape guard and
carries a pointer to where the pair went. All three mutations still red from the
new home: pasting a seven-word rubric run into the brief, truncating the
altitude tell in its home, and re-quoting the tell in the brief.

Worth recording: this is a gate-set blind spot, not a knowledge gap. `make
lint-ruff` and the pack suite were green throughout; only `build-check`'s
structural boundary check sees it, and this session had not run build-check
since before the guards were written.
