+++ packs/core/tests/skills/new-spec/test_acceptance_criteria_discipline.py
@@ -11,8 +11,6 @@ SPEC = PACK_ROOT / ".apm/skills/new-spec/assets/spec.md"
 PLAN = PACK_ROOT / ".apm/skills/new-spec/assets/plan.md"
 EVALS = PACK_ROOT / ".apm/skills/new-spec/evals/evals.json"
 RUBRIC = PACK_ROOT / ".apm/skills/new-spec/references/spec-authoring-rubric.md"
-# The brief quotes the rubric's altitude tell; the pair is pinned below.
-BRIEF = PACK_ROOT.parents[1] / "docs/product/briefs/agent-authoring-input-quality.md"
 
 # The rubric joins SOURCES so every pinned rule below also asserts its absence
 # there. The rubric points at the owning surface for criterion shape; a future
@@ -408,90 +406,11 @@ def test_rubric_defers_criterion_shape_and_stays_authoring_guidance() -> None:
     assert "a ceiling and a stall point, never a floor" in text
 
 
-ALTITUDE_TELL = (
-    "a design position, an evidence base, an inventory, or a governance concern "
-    "rather than citing one"
-)
-
-
-def test_altitude_tell_lives_whole_in_the_rubric_and_nowhere_else() -> None:
-    """The four-clause tell must survive its re-homing, and stay single-homed.
-
-    Moving the three altitude tells out of the brief and into the rubric
-    silently dropped the fourth clause, "or a governance concern", and left the
-    brief quoting a third wording that matched neither. A phrase pin on the
-    rubric alone did not notice, because a clause can go missing from the
-    middle of the phrase it pins.
-
-    So pin both halves of the contract the owning brief now declares: the whole
-    clause list is present in the rubric, and the brief describes no part of the
-    rubric's content, so the tell text must not appear there at all. Truncating
-    the rubric reds the first assertion; pasting the tell back into the brief
-    reds the second.
-    """
-    assert ALTITUDE_TELL in flattened(RUBRIC), (
-        "the altitude tell lost a clause in its own home"
-    )
-    assert ALTITUDE_TELL not in flattened(BRIEF), (
-        "the brief restates the rubric's tell; it must cite the class, not its text"
-    )
-
-
-# The owning brief declares that it describes no part of the rubric's content.
-#
-# Calibration, measured 2026-09-08 after the restatements were cut, excluding
-# Markdown table delimiters (`| --- | --- |` runs match trivially and carry no
-# prose): the two files share 2 six-word runs and 0 of seven or more. Both are
-# `an implementation loop with gates between`, incidental shared phrasing about
-# the same benchmark caveat, and `authored where an owner already exists,`,
-# class 1's subject named in the brief's Outcome. Neither extends to seven in
-# either direction, so seven is the shortest run length that reds only on
-# restatement.
-#
-# One structural caveat the threshold does not cover: a rubric *heading* of
-# seven words or more, quoted by name in the brief, would red this guard on
-# correct text. An earlier revision had exactly that case and it disappeared
-# when the section was renamed, which is why it is recorded rather than assumed
-# absent.
-#
-# An earlier revision set this to eight while the two substantive restatements
-# were seven words long — a threshold calibrated one word above the duplication
-# it existed to catch, which is why the number and its evidence are recorded
-# here rather than asserted.
-RESTATEMENT_RUN_WORDS = 7
-
-
-def test_the_brief_restates_no_run_of_the_rubrics_text() -> None:
-    """Enforce the cut, not just perform it.
-
-    Six review rounds produced eleven findings about which document owned or
-    described which rule, because each repair moved text and left a sentence
-    describing where it went. The brief now declares it describes no part of the
-    rubric's content; this makes that declaration checkable instead of another
-    claim that can go stale.
-
-    Named blind spots: a paraphrase, or any verbatim run shorter than the
-    threshold, passes — two six-word runs are admitted today and recorded in the
-    calibration note above. A restatement in any file other than these two also
-    passes. This catches verbatim drift between the declared owner and its
-    brief, which is the failure that actually recurred.
-    """
-    rubric_words = flattened(RUBRIC).split()
-    brief = flattened(BRIEF)
-    n = RESTATEMENT_RUN_WORDS
-    shared = sorted(
-        {
-            run
-            for i in range(len(rubric_words) - n + 1)
-            # Table delimiters are structure, not prose.
-            if "---" not in (run := " ".join(rubric_words[i : i + n]))
-            and run in brief
-        }
-    )
-    assert not shared, (
-        f"the brief restates {len(shared)} run(s) of {n}+ words from the rubric; "
-        f"cite the class by number instead. First: {shared[0]!r}"
-    )
+# The altitude-tell single-homing guard and the brief-versus-rubric overlap
+# guard live in `tests/roster/test_spec_authoring_rubric_brief_boundary.py`:
+# both relate this pack to a `docs/product/` brief, so they cross the pack
+# boundary that `pack-tests-stay-in-pack` enforces. Everything below stays
+# pack-local.
 
 
 def test_rubric_ships_derivations_and_cites_no_internal_locator() -> None:
